backend/requirements.txt
# GUARDZILLA Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# AI/ML Models
ultralytics==8.1.0
torch==2.0.1
torchvision==0.15.2
face-recognition==1.3.5
insightface==0.7.3

# Tracking
deep-sort-realtime==1.3.2

# Database
sqlite3

# Utilities
pyyaml==6.0.1
requests==2.31.0
aiofiles==23.2.1

# Async & Scheduling
APScheduler==3.10.4
asyncio-contextmanager==1.0.0

# Hardware (Raspberry Pi)
RPi.GPIO==0.7.0
gpiozero==2.0.1

# Email
secure-smtplib==0.1.1

# Logging & Monitoring
python-logging-loki==0.3.2

# Development
python-dotenv==1.0.0
